added CAN FD support

+++ include/lwcan/buffer.h
@@ -15,20 +15,20 @@ struct lwcan_buffer
 
     uint8_t *payload;
 
-    uint16_t length;
+    uint32_t length;
 };
 
-struct lwcan_buffer *lwcan_buffer_new(uint16_t length);
+struct lwcan_buffer *lwcan_buffer_new(uint32_t length);
 
 void lwcan_buffer_delete(struct lwcan_buffer *buffer);
 
-lwcanerr_t lwcan_buffer_copy_to(struct lwcan_buffer *buffer, const uint8_t *source, uint16_t length);
+lwcanerr_t lwcan_buffer_copy_to(struct lwcan_buffer *buffer, const uint8_t *source, uint32_t length);
 
-lwcanerr_t lwcan_buffer_copy_to_offset(struct lwcan_buffer *buffer, const uint8_t *source, uint16_t length, uint16_t offset);
+lwcanerr_t lwcan_buffer_copy_to_offset(struct lwcan_buffer *buffer, const uint8_t *source, uint32_t length, uint32_t offset);
 
-lwcanerr_t lwcan_buffer_copy_from(struct lwcan_buffer *buffer, uint8_t *distination, uint16_t length);
+lwcanerr_t lwcan_buffer_copy_from(struct lwcan_buffer *buffer, uint8_t *distination, uint32_t length);
 
-lwcanerr_t lwcan_buffer_copy_from_offset(struct lwcan_buffer *buffer, uint8_t *distination, uint16_t length, uint16_t offset);
+lwcanerr_t lwcan_buffer_copy_from_offset(struct lwcan_buffer *buffer, uint8_t *distination, uint32_t length, uint32_t offset);
 
 #ifdef __cplusplus
 }
